Update convexHullGrahamScan.py
Cleaned up code. Sent copies of arrays rather than the array itself to certain functions which removed minor errors/data changes. Lines 132-135 are now purely computation and 137-141 are outputs for the user making it a little easier to follow. Also added time to show the time it takes to complete the entire code (~0.001s average).